extract function node logic into helpers
